feat(review): add course reviews

+++ backend/src/main/java/br/com/tecflix_app/data/DTO/v1/response/ReviewDTO.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 @Setter @Getter
 public class ReviewDTO {
     private Long id;
-    private Double score;
+    private Integer score;
     private String comment;
     private LocalDateTime createdAt;
     private UserDTO user;
