+++ backend/src/main/resources/db/migration/structure/V10__create_table_reviews.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS reviews(
     id SERIAL PRIMARY KEY,
     user_id UUID REFERENCES users(id) NOT NULL,
     course_id UUID REFERENCES courses(id) NOT NULL,
-    score FLOAT NOT NULL,
+    score INTEGER NOT NULL,
     comment VARCHAR(255),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL
 );
